requirements.txt
numpy~=1.23
gymnasium>=0.27.1
pygame>=2.3.0
pettingzoo>=1.22.3
matplotlib==3.8.4
pytest>=8.0.0
numba>=0.59.0
wheel
GDAL==3.4.1
opendrift==1.10.7
h5py
xarray
copernicusmarine>=2.1.0
python-dotenv